backend/requirements.txt
Django==3.2.8
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
PyJWT==1.7.1
django-cors-headers==3.10.0
gunicorn==20.1.0
psycopg2==2.9.1
whitenoise==5.3.0
